+++ source/MRViewer/ImGuiMenu.cpp
@@ -3169,19 +3169,4 @@ bool ImGuiMenu::UiRenderManagerImpl::canConsumeEvent( BasicUiRenderTask::Interac
         bool( consumedInteractions & event );
 }
 
-void showModal( const std::string& msg, NotificationType type )
-{
-    if ( auto menu = getViewerInstance().getMenuPlugin() )
-        menu->showModalMessage( msg, type );
-    else
-    {
-        if ( type == NotificationType::Error )
-            spdlog::error( "Show Error: {}", msg );
-        else if ( type == NotificationType::Warning )
-            spdlog::warn( "Show Warning: {}", msg );
-        else //if ( type == MessageType::Info )
-            spdlog::info( "Show Info: {}", msg );
-    }
-}
-
 } // end namespace
