+++ source/MRViewer/MRSceneObjectsListDrawer.cpp
@@ -1,20 +1,21 @@
 #include "MRSceneObjectsListDrawer.h"
 #include "MRSceneCache.h"
-#include "MRMesh/MRObject.h"
-#include "MRMesh/MRObjectsAccess.h"
 #include "MRViewer.h"
 #include "MRViewerInstance.h"
 #include "MRViewport.h"
 #include "MRUIStyle.h"
-#include "MRMesh/MRSceneRoot.h"
+#include "MRShowModal.h"
 #include "MRRibbonConstants.h"
 #include "MRColorTheme.h"
 #include "ImGuiMenu.h"
-#include "MRMesh/MRChangeSceneAction.h"
-#include "MRMesh/MRChangeXfAction.h"
 #include "MRAppendHistory.h"
 #include "MRRibbonSchema.h"
 #include "MRUITestEngine.h"
+#include "MRMesh/MRObject.h"
+#include "MRMesh/MRObjectsAccess.h"
+#include "MRMesh/MRSceneRoot.h"
+#include "MRMesh/MRChangeSceneAction.h"
+#include "MRMesh/MRChangeXfAction.h"
 #include "MRPch/MRSpdlog.h"
 #include "imgui_internal.h"
 #include "imgui.h"
