+++ source/MRViewer/MRViewer.vcxproj
@@ -50,6 +50,7 @@
     <ClCompile Include="MRObjectPointsHistory.cpp" />
     <ClCompile Include="MRRibbonMenuSearch.cpp" />
     <ClCompile Include="MRSceneTextureGL.cpp" />
+    <ClCompile Include="MRShowModal.cpp" />
     <ClCompile Include="MRSurfaceContoursWidget.cpp" />
     <ClCompile Include="MRToolsLibrary.cpp" />
     <ClCompile Include="MRLinesShader.cpp" />
@@ -178,6 +179,7 @@
     <ClInclude Include="MRObjectPointsHistory.h" />
     <ClInclude Include="MRRibbonMenuSearch.h" />
     <ClInclude Include="MRSceneTextureGL.h" />
+    <ClInclude Include="MRShowModal.h" />
     <ClInclude Include="MRSpaceMouseParameters.h" />
     <ClInclude Include="MRSurfaceContoursWidget.h" />
     <ClInclude Include="MRToolsLibrary.h" />
