+++ source/MRViewer/MRViewer.vcxproj.filters
@@ -436,6 +436,9 @@
     <ClCompile Include="MRSceneOperations.cpp">
       <Filter>Helpers</Filter>
     </ClCompile>
+    <ClCompile Include="MRShowModal.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ImGuiHelpers.h">
@@ -850,6 +853,9 @@
     </ClInclude>
     <ClInclude Include="MRSceneOperations.h">
       <Filter>Helpers</Filter>
+    </ClInclude>
+    <ClInclude Include="MRShowModal.h">
+      <Filter>Source Files</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
