+++ source/MRViewer/MRViewerSettingsPlugin.cpp
@@ -1,31 +1,32 @@
 #include "MRViewerSettingsPlugin.h"
-#include "MRViewer/MRRibbonMenu.h"
-#include "MRViewer/ImGuiHelpers.h"
-#include "MRViewer/MRColorTheme.h"
-#include "MRViewer/ImGuiHelpers.h"
-#include "MRViewer/MRMouseController.h"
-#include "MRViewer/MRViewport.h"
+#include "MRRibbonMenu.h"
+#include "ImGuiHelpers.h"
+#include "MRColorTheme.h"
+#include "ImGuiHelpers.h"
+#include "MRMouseController.h"
+#include "MRViewport.h"
 #include "MRMesh/MRObjectsAccess.h"
-#include "MRViewer/MRCommandLoop.h"
-#include "MRViewer/MRViewerSettingsManager.h"
-#include "MRViewer/MRGLMacro.h"
-#include "MRViewer/MRGladGlfw.h"
-#include "MRViewer/MRRibbonConstants.h"
-#include "MRViewer/MRViewer.h"
-#include "MRViewer/MRImGuiVectorOperators.h"
+#include "MRCommandLoop.h"
+#include "MRViewerSettingsManager.h"
+#include "MRGLMacro.h"
+#include "MRGladGlfw.h"
+#include "MRRibbonConstants.h"
+#include "MRViewer.h"
+#include "MRImGuiVectorOperators.h"
 #include "MRMesh/MRSystem.h"
-#include "MRViewer/MRSpaceMouseHandlerHidapi.h"
+#include "MRSpaceMouseHandlerHidapi.h"
 #include "MRMesh/MRLog.h"
 #include "MRPch/MRSpdlog.h"
-#include "MRViewer/MRUIStyle.h"
+#include "MRUIStyle.h"
 #include "MRMesh/MRStringConvert.h"
 #include "MRMesh/MRSceneSettings.h"
 #include "MRMesh/MRDirectory.h"
 #include <MRMesh/MRSceneRoot.h>
-#include <MRViewer/MRFileDialog.h>
+#include <MRFileDialog.h>
 #include "MRMesh/MRObjectMesh.h"
-#include "MRViewer/MRRibbonSceneObjectsListDrawer.h"
-#include "MRViewer/MRUnitSettings.h"
+#include "MRRibbonSceneObjectsListDrawer.h"
+#include "MRUnitSettings.h"
+#include "MRShowModal.h"
 
 namespace
 {
